# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Classification": {
        "enum": [
          "PUBLIC",
          "INTERNAL",
          "CONFIDENTIAL",
          "RESTRICTED"
        ],
        "title": "Classification",
        "type": "string"
      },
...
      "CreateEvaluationCaseRequest": {
        "properties": {
          "expected": {
            "additionalProperties": true,
            "title": "Expected",
            "type": "object"
          },
          "external_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "External Id",
            "type": "string"
          },
          "fixtures": {
            "additionalProperties": true,
            "title": "Fixtures",
            "type": "object"
          },
          "input_payload": {
            "additionalProperties": true,
            "title": "Input Payload",
            "type": "object"
          },
          "version": {
            "default": 1,
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "external_id",
          "input_payload",
          "expected"
        ],
        "title": "CreateEvaluationCaseRequest",
        "type": "object"
      },
      "CreateEvaluationDatasetRequest": {
        "properties": {
          "description": {
            "default": "",
            "maxLength": 4000,
            "title": "Description",
            "type": "string"
          },
          "domain": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Domain",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "domain"
        ],
        "title": "CreateEvaluationDatasetRequest",
        "type": "object"
      },
...
      "EvaluationCaseView": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dataset_id": {
            "format": "uuid",
            "title": "Dataset Id",
            "type": "string"
          },
          "expected": {
            "additionalProperties": true,
            "title": "Expected",
            "type": "object"
          },
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "fixtures": {
            "additionalProperties": true,
            "title": "Fixtures",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "input_payload": {
            "additionalProperties": true,
            "title": "Input Payload",
            "type": "object"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "dataset_id",
          "external_id",
          "version",
          "input_payload",
          "expected",
          "fixtures",
          "created_at"
        ],
        "title": "EvaluationCaseView",
        "type": "object"
      },
      "EvaluationDatasetView": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "domain",
          "created_at",
          "updated_at"
        ],
        "title": "EvaluationDatasetView",
        "type": "object"
      },
      "EvaluationRunView": {
        "properties": {
          "agent_version_id": {
            "format": "uuid",
            "title": "Agent Version Id",
            "type": "string"
          },
          "application_revision": {
            "title": "Application Revision",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dataset_id": {
            "format": "uuid",
            "title": "Dataset Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "metrics": {
            "additionalProperties": true,
            "title": "Metrics",
            "type": "object"
          },
          "model_profile_id": {
            "format": "uuid",
            "title": "Model Profile Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "dataset_id",
          "agent_version_id",
          "model_profile_id",
          "application_revision",
          "status",
          "metrics",
          "created_at",
          "updated_at"
        ],
        "title": "EvaluationRunView",
        "type": "object"
      },
      "EventView": {
        "properties": {
          "causation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Causation Id"
          },
          "classification": {
            "$ref": "#/components/schemas/Classification"
          },
          "correlation_id": {
            "format": "uuid",
            "title": "Correlation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "integer"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "sequence",
          "name",
          "run_id",
          "causation_id",
          "correlation_id",
          "schema_version",
          "classification",
          "payload",
          "created_at"
        ],
        "title": "EventView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StartEvaluationRunRequest": {
        "properties": {
          "agent_version_id": {
            "format": "uuid",
            "title": "Agent Version Id",
            "type": "string"
          },
          "application_revision": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Application Revision",
            "type": "string"
          },
          "model_profile_id": {
            "format": "uuid",
            "title": "Model Profile Id",
            "type": "string"
          }
        },
        "required": [
          "agent_version_id",
          "model_profile_id",
          "application_revision"
        ],
        "title": "StartEvaluationRunRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Enterprise Agent Runtime and Intelligence Workspace",
    "title": "Obsion API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/evaluations/datasets": {
      "get": {
        "operationId": "list_datasets_api_v1_admin_evaluations_datasets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvaluationDatasetView"
                  },
                  "title": "Response List Datasets Api V1 Admin Evaluations Datasets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Datasets",
        "tags": [
          "administration",
          "evaluations"
        ]
      },
      "post": {
        "operationId": "create_dataset_api_v1_admin_evaluations_datasets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvaluationDatasetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationDatasetView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Dataset",
        "tags": [
          "administration",
          "evaluations"
        ]
      }
    },
    "/api/v1/admin/evaluations/datasets/{dataset_id}/cases": {
      "get": {
        "operationId": "list_cases_api_v1_admin_evaluations_datasets__dataset_id__cases_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvaluationCaseView"
                  },
                  "title": "Response List Cases Api V1 Admin Evaluations Datasets  Dataset Id  Cases Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Cases",
        "tags": [
          "administration",
          "evaluations"
        ]
      },
      "post": {
        "operationId": "add_case_api_v1_admin_evaluations_datasets__dataset_id__cases_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvaluationCaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationCaseView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Case",
        "tags": [
          "administration",
          "evaluations"
        ]
      }
    },
    "/api/v1/admin/evaluations/datasets/{dataset_id}/runs": {
      "post": {
        "operationId": "run_evaluation_api_v1_admin_evaluations_datasets__dataset_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartEvaluationRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run Evaluation",
        "tags": [
          "administration",
          "evaluations"
        ]
      }
    },
    "/api/v1/admin/evaluations/runs": {
      "get": {
        "operationId": "list_runs_api_v1_admin_evaluations_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "dataset_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvaluationRunView"
                  },
                  "title": "Response List Runs Api V1 Admin Evaluations Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Runs",
        "tags": [
          "administration",
          "evaluations"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/events": {
      "get": {
        "operationId": "list_run_events_api_v1_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 2000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventView"
                  },
                  "title": "Response List Run Events Api V1 Runs  Run Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Run Events",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/runs/{run_id}/events/stream": {
      "get": {
        "operationId": "stream_run_events_api_v1_runs__run_id__events_stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stream Run Events",
        "tags": [
          "events"
        ]
      }
    },
...
  }
}